Added version 5.4.2753

+++ chromium/chrome/browser/downgrade/snapshot_file_collector.cc
@@ -27,7 +27,7 @@ namespace downgrade {
 
 SnapshotItemDetails::SnapshotItemDetails(base::FilePath path,
                                          ItemType item_type,
-                                         int data_types,
+                                         uint64_t data_types,
                                          SnapshotItemId id)
     : path(std::move(path)),
       is_directory(item_type == ItemType::kDirectory),
@@ -60,7 +60,7 @@ std::vector<SnapshotItemDetails> CollectProfileItems() {
   // kPrefLastLaunchTime and kUserTriggeredIsolatedOrigins prefs have to be
   // reset. When data type content is deleted, blocklisted sites are deleted
   // from the translation prefs.
-  int pref_data_type =
+  uint64_t pref_data_type =
       content::BrowsingDataRemover::DATA_TYPE_COOKIES |
       chrome_browsing_data_remover::DATA_TYPE_ISOLATED_ORIGINS |
       chrome_browsing_data_remover::DATA_TYPE_HISTORY |
